Move japicmp from build.gradle to dedicated plugin
Prior to this commit, the japicmp Gradle configuration would be declared
in the main build.gradle file. Whenever some API needed to be removed,
contributors first needed to declare that API in the configuration as an
exclude. This approach makes it harder to remove deprecated APIs.

This commit moves the japicmp configuration from the build script to a
dedicated custom plugin in buildSrc. This plugin, when applied to a
project, configures a `apiCompatibility` task that checks for API
compatibility issues with the baseline version but silently ignores
removals for `@Deprecated` methods, fields and classes.